feat: Cli argument open now provides same functionality as before
Moved logic to open new or switch to existing session into `open`
module.

Refs: #2
