+++ src/commands/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod cli;
+pub(crate) mod open;
